+++ Changelog.md
@@ -1,8 +1,24 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.3.4
+Added an option to set a custom texture for the absorb/healcomm on the unitframe  
+Added options to enable the static backdrop size in Nameplates  
+Added an option to flip the target unitframe backdrop texture  
+Added an option to disable shadows in the OmniCD skin  
+Added portraits to the Boss unitframe
+Added Russian locale by ZamestoTV  
+Added gradient to Boss Castbars  
+Updated Portraits function to enable the required 3D model when enabling the portrait  
+Updated profile to remove some references to warlock class resource colors  
+Updated the chat Race icon to be disabled by default  
+Updated German locale  
+Updated for 11.1.7  
+Fixed an issue where classbar shadows would not work when detached  
+___
 ### 4.3.3.1
 Fixed an issue where chat could error  
+___
 ### 4.3.3
 Added enableMouseoverCast CVar to the cvar list due to it being needed for mouse over casts with Clique  
 Added OverwatchHQ, OverwatchLQ, OverwatchLQ2 and OverwatchLQFlip statusbar textures